move vmdpy
