+++ sktime/transformations/series/tests/test_vmd.py
@@ -10,8 +10,8 @@ import pytest
 from sktime.forecasting.compose import TransformedTargetForecaster
 from sktime.forecasting.trend import TrendForecaster
 from sktime.tests.test_switch import run_test_for_class
-from sktime.transformations.series.vmd._vmd import VmdTransformer
-from sktime.transformations.series.vmd._vmdpy import VMD
+from sktime.transformations.series.vmd import VmdTransformer
+from sktime.libs.vmdpy import VMD
 
 
 def _generate_vmd_testdata(T=1000, f_1=2, f_2=24, f_3=288, noise=0.1):
