--- sktime/transformations/series/vmd/__init__.py
@@ -1,9 +1,0 @@
-# copyright: sktime developers, BSD-3-Clause License (see LICENSE file)
-"""VMD transformer and VMD mini-package based on pyvmd."""
-
-__author__ = ["DaneLyttinen", "vcarvo"]
-
-
-from sktime.transformations.series.vmd._vmd import VmdTransformer
-
-__all__ = ["VmdTransformer"]
