--- sktime/transformations/series/vmd/tests/__init__.py
@@ -1,2 +1,0 @@
-# copyright: sktime developers, BSD-3-Clause License (see LICENSE file)
-"""Tests for VMD transformer and VMD mini-package based on pyvmd."""
